backend/requirements.txt
Django===4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter>=22.0
python-decouple==3.8
PyJWT==2.12.1
Pillow>=12.0.0
psycopg2-binary>=2.9.0
